misc: (is_const_classvar) move to utils
